Further cleanup.

+++ utility/streams/stream.h
@@ -4,6 +4,8 @@
 namespace utility {
 	class stream {};
 
+	// base stream writer
+
 	template<typename type, typename stream_type>
 	struct stream_writer {
 		static void write(const type& value) {
@@ -201,8 +203,8 @@ namespace utility {
 	};
 
 	template<typename stream_type>
-	struct stream_writer<f64, stream_type> {
-		static void write(f64 value) {
+	struct stream_writer<f32, stream_type> {
+		static void write(f32 value) {
 			// handle negative values
 			if(value < 0) {
 				value = -value;
@@ -211,11 +213,11 @@ namespace utility {
 			// extract the exponent
 			i32 exponent = 0;
 
-			if(value != 0.0) {
+			if(value != 0.0f) {
 				exponent = static_cast<i32>(std::floor(std::log10(value)));
 
 				if(exponent < -3 || exponent > 3) {
-					value /= std::pow(10.0, exponent);
+					value /= std::pow(10.0f, exponent);
 				}
 				else {
 					exponent = 0;
@@ -224,21 +226,21 @@ namespace utility {
 
 			// convert the mantissa to a string
 			i32 int_part = static_cast<i32>(value);
-			f64 fraction = value - int_part;
+			f32 fraction = value - static_cast<f32>(int_part);
 			static char buffer[100];
 			u8 index = 0;
 
 			stream_writer<i32, stream_type>::write(int_part);
 
 			// fractional part
-			if(fraction > 0.0) {
+			if(fraction > 0.0f) {
 				buffer[index++] = '.';
 
 				for(u8 i = 0; i < 4; i++) {
 					fraction *= 10;
 					const u8 digit = static_cast<u8>(fraction);
 					buffer[index++] = static_cast<char>(digit + 48);
-					fraction -= digit;
+					fraction -= static_cast<f32>(digit);
 				}
 
 				// remove trailing zeros
@@ -279,8 +281,8 @@ namespace utility {
 	};
 
 	template<typename stream_type>
-	struct stream_writer<f32, stream_type> {
-		static void write(f32 value) {
+	struct stream_writer<f64, stream_type> {
+		static void write(f64 value) {
 			// handle negative values
 			if(value < 0) {
 				value = -value;
@@ -289,11 +291,11 @@ namespace utility {
 			// extract the exponent
 			i32 exponent = 0;
 
-			if(value != 0.0f) {
+			if(value != 0.0) {
 				exponent = static_cast<i32>(std::floor(std::log10(value)));
 
 				if(exponent < -3 || exponent > 3) {
-					value /= std::pow(10.0f, exponent);
+					value /= std::pow(10.0, exponent);
 				}
 				else {
 					exponent = 0;
@@ -302,21 +304,21 @@ namespace utility {
 
 			// convert the mantissa to a string
 			i32 int_part = static_cast<i32>(value);
-			f32 fraction = value - static_cast<f32>(int_part);
+			f64 fraction = value - int_part;
 			static char buffer[100];
 			u8 index = 0;
 
 			stream_writer<i32, stream_type>::write(int_part);
 
 			// fractional part
-			if(fraction > 0.0f) {
+			if(fraction > 0.0) {
 				buffer[index++] = '.';
 
 				for(u8 i = 0; i < 4; i++) {
 					fraction *= 10;
 					const u8 digit = static_cast<u8>(fraction);
 					buffer[index++] = static_cast<char>(digit + 48);
-					fraction -= static_cast<f32>(digit);
+					fraction -= digit;
 				}
 
 				// remove trailing zeros
